emv: Move testFindDefiniteReallyLong to Android tests to work around logging issue.
